+++ server/package.json
@@ -30,7 +30,8 @@
     "@types/node": "^18.7.15",
     "concurrently": "^7.3.0",
     "nodemon": "^2.0.19",
-    "typescript": "^4.7.4"
+    "typescript": "^4.7.4",
+    "@types/ws": "^8.5.3"
   },
   "type": "module"
 }
