+++ server/util.ts
@@ -1,0 +1,13 @@
+import * as dotenv from "dotenv";
+import jwt from "jsonwebtoken";
+dotenv.config();
+
+export function verifyLogin(token: string | undefined) {
+    try {
+        if (token && jwt.verify(token, process.env.jwt_secret + "", { algorithms: ["HS256"] })) {
+            return true;
+        }
+    } catch(err) {
+        return false;
+    }
+}
